Update 1.3
Commenting methods.

Creating new instance variable FileStatus for identifying whether the
file is successfully read or not.

Using TreeSet to store keywords and contentWords, as they will not be
duplicate, plus TreeSet provides good performance in terms of accessing
and searching the collection.

Successfully extracting the keywords and content words.