Implemented parseInfixForm()
A number of methods that support parsing the Infix form of query are
added, where the parseInfixString(String str) and the
parseCoveringBracket(String q) are the recursive methods.

The parseAndInfix(String left, String right), parseOrInfix(String left,
String right), and the parseNotInfix(String right) are called to process
the queries, they will call back to the parseInfixString(String str)
recursively until the query is atomic or a NotQuery.