Fixed Incorrect Logic
Fixed the incorrect logical design of Queries. NotQuery.matches() should
return all the unmatched results. OrQuery.matches() should not ignore
the NotQuery. AndQuery.matches() should treat all the subqueries as the
same.

WebIndex is updated; a new method-getWebDocsMap()- is added that returns
the deep copy of the hashMap stored in the webindex.