Modified SearchActionHandler, updated ConfirmActionHandler for history
The query is now in lowercase and trimmed.It is now like this:
query = infixQuery.getText().trim().toLowerCase();
The demonstration of query result is optimised.

A new set is added for managing the history. (browsingHistory). The
collection in controller class that was created for webDocs is deleted.
(webDocCollection) the ConfirmActionHandler class is updated for
handling the browsing history