Automatic tab switching and navigation
Every time a new tab is created, the name of each tab is assigned based
on the number of tabs exist (e.g., Tab1, Tab2, Tab3 etc.), the
application also navigates to/ shows the last tab created automatically.

Further, when user searches using query and click on the results of the
query, the application navigates to the tab that is created for the
'result' clicked.