New: Added a windowListener to the frame
This windowListener is used to save the index and history in a file
called log.txt when the user closes the application.

When the user execute the application, the applications will pop up a
dialogue to ask whether the  user wants to read the previous history. If
yes, the application will load all the history including the index
previously created.