Implement MenuButton
Previous Button was used to represent the "Menu" in this program for switching the view between the DisplayPane and the QueryPane. Both "views" need to have this Menu, so this class implements this feature for both views to use directly.

new file:   IndexingWebBrowser/src/main/java/com/curtisnewbie/view/MenuButton.java