Complete reimplementation of WebDoc
This is based on the previous refactoring of WebDoc, however, Jsoup is used within WebDoc, so that each WebDoc represents one single webpage, and the head and body of the webpage is seperated within WebDoc.

Irrelavant methods that are not needed in this program are removed.

modified:   IndexingWebBrowser/src/main/java/com/curtisnewbie/webBrowserModel/WebDoc.java