Update Url TextField when the webpage is successfully loaded
update textField to tell the user the location of current webpage (just like normal browser does). If it is a online webpage, the textfield will show the location of the url, if it is a local webpage, it will show the original location that the user entered.

modified:   IndexingWebBrowser/src/main/java/com/curtisnewbie/controller/BrowserController.java