Fix bug in updating the QueryResultPanel
The bug occurs when user enters a query that doesn't match anything in the WebIndex for both head and body, the QueryResultPanel doesn't update itself at all but rather throws a NullPointerException.

For some reasons, it seems that while modifying the ObservableList of a Node (here, it's VBox) that is being displayed, the ObservableList should not be obtained and stored as a variable. Rather, we should call a method directly on it.

E.g.,
	someVBox.getChildren().clear();
	someVBox.getChildren().add(new Button());

modified:   IndexingWebBrowser/src/main/java/com/curtisnewbie/controller/BrowserController.java