Fix bug in going backward and forward in history
Clicking the backTrackBtn or forwardBtn in displayPane allows going backward and forward in history in the current tab. However, when there is no more histories that allows to go backward or forward, IndexOutOfBoundException is being thrown. This is caused by the incorrect checking of index, and it's now fixed.

modified:   IndexingWebBrowser/src/main/java/com/curtisnewbie/controller/BrowserController.java