Optimise QueryBuilder.parse() method
It shouldn't check whether the query contains any of the three operators. Atomic Query should be found as long as the query doesn't start with "and", "or", "not". Since it method is parsing prefix query, it will be quite consistent already.

This method is also changed in that only return statement is used. Using multiple return statement may be problematic. Variable "Starting_Index" is removed, it might be better to define it directly in the method parameters to improve readability.

JavaDoc is also updated.

modified:   src/main/java/com/curtisnewbie/webBrowserModel/QueryBuilder.java