Remove instance variable (queryResults) in AndQuery class
ArrayList<Set<WebDoc>> queryResults shouldn't be instantiated as an instance variable, else it should be re-instantiated everytime the matches() method is called. Without re-instantiating this variable, the result sets in this List will be increased everytime the matches() method is called eventhough the provided WebIndex is the same. Which is non-sense.

modified:   IndexingWebBrowser/src/main/java/com/curtisnewbie/webBrowserModel/AndQuery.java